Drop support for Rails < 6.0, Ruby < 2.7.
Support for Rails 5.2 and Ruby 2.6 ended in 2018 so I think it's a reasonable time to
drop support for it and everything below it.

I've changed the Rails version check in `application.rb` because `represent_boolean_as_integer` was
deprecated in Rails 6.0[^1].

The target Ruby version in `.rubocop.yml` has been updated but it seems
like Rubocop isn't actually being used to autocorrect offenses in the
gem and I didn't want to autocorrect everything because there are
hundreds of offenses.

[^1]:(https://github.com/rails/rails/blob/e29f5e0feb68ee6a9bd44f24991d7494a71884f0/guides/source/6_0_release_notes.md?plain=1#L446)